found the root of the zombie shell issue
On termination the bash shell object would clean it self up so it could
be reused with a fresh init. However, shells that had been started with
a trap e.g. like MTM-SSH, would run into an issue where the base shell
would terminate on ssh exiting, but because bash cleaned itself up, the
child could attempt sending more commands, which spawned a new base
shell, however this new base shell would not have the guard process, as
it terminated with the first base shell.

If the new shell was not shut down cleanly, e.g. php timeout, runtime
exception whatever the new base shell would keep on living for 1000 days
(limit of the sleep that keeps the stdIn pipe open).

The change adds a boolean class attribute to termination, when set the
shell cannot be reused it is forever dead.

In addition the termination logic got a make over to deal with various
edge cases where a terminating shell was in a dirty state e.g. the user
issued an "exit" command outside the terminate() function. The shell
will now exit clean even if the state is not as expected.